+++ src/components/Main/Modal/ReportModal/ReportModal.tsx
@@ -17,6 +17,7 @@ import Swal from 'sweetalert2';
 import { toast } from 'react-toastify';
 import { AxiosError } from 'axios';
 import { ErrorData } from '../../DocumentForm/commonInterface';
+import DatePickerComponent from '../../DocumentForm/components/DatePicker/DatePickerComponent';
 export type ReportInfo = {
   title: string;
   body: string;
@@ -218,30 +219,3 @@ const ReportModal = (value: ReportModalProps) => {
 };
 
 export default ReportModal;
-
-interface DatePickerProps {
-  setPropsDate: React.Dispatch<React.SetStateAction<Date | undefined>>;
-}
-
-const DatePickerComponent = (props: DatePickerProps) => {
-  const [startDate, setStartDate] = useState(new Date());
-  const changeDateHandler = (date: Date) => {
-    date !== null && setStartDate(date);
-    props.setPropsDate(date);
-  };
-  return (
-    <StDatePicker
-      selected={startDate}
-      onChange={(date: Date) => changeDateHandler(date)}
-    />
-  );
-};
-
-const StDatePicker = styled(DatePicker)`
-  box-sizing: border-box;
-  border: none;
-  width: 150px;
-  box-shadow: rgba(0, 0, 0, 0.16) 0px 1px 4px;
-  height: 30px;
-  padding-left: 15px;
-`;
